fixed nulls

+++ Magic_Astronauts/Core/Weather.cs
@@ -5,6 +5,6 @@ public class Weather
     public int WeatherID { get; set; }
     public DateTime Date { get; set; }
     public string Location { get; set; }
-    public float Temp { get; set; }
+    public decimal Temp { get; set; }
     public int Humidity { get; set; }
 }
